Begin cleaning up websocket-based useEffect (...) - Extract socket update functionality to new method. - Rename interval for clarity. - Extract initial illumination logic to separate util function. - Remove redundant useEffect hook, and replace with simple ternary operator when accounting for presence of color temperature.
